Move the three.js-dependent constants out of viewer/constants
viewer/constants exported OrthoBaseRotations and OrthoCamerasBaseRotations, both built with
`new Euler(...)` at module scope. Module-scope construction means three.js can never be
tree-shaken out of anything that touches the module - and viewer/constants is imported by
331 modules, including navbar.tsx and the web workers.

The cost landed squarely on async_bucket_picker.worker, which does nothing but pick buckets
to load: 272.77 kB, of which ~215 kB was three.js reached solely through this one import.
Every other module in that worker's graph is three-free, so moving these two constants into
viewer/ortho_base_rotations removes three.js from it entirely.

  async_bucket_picker.worker: 272.77 kB -> 51.76 kB (-81%, three.js: 0 modules remaining)

Only four non-test modules use these constants - scene_controller, camera_controller,
skeleton_handlers and skeletontracing_saga - and all four already import three themselves,
so the move costs them nothing. `yarn check-cyclic-dependencies` passes: the new module
imports OrthoViews from viewer/constants, and nothing imports back.

Note this does not change page load. three.js stays in the entry chunk, because
route_wrappers statically imports the viewer, so the entry is unchanged at ~6,545 kB
(1,838 kB gzip). Evicting three.js from the initial payload would additionally require the
viewer to be lazily loaded, which was deliberately abandoned in #9991. This is purely a
worker-size fix.
