+++ frontend/javascripts/viewer/controller/combinations/skeleton_handlers.ts
@@ -5,7 +5,7 @@ import pick from "lodash-es/pick";
 import { Euler, Matrix4, Scene, Vector3 as ThreeVector3 } from "three";
 import type { AdditionalCoordinate } from "types/api_types";
 import type { OrthoView, Point2, Vector3, Viewport } from "viewer/constants";
-import { OrthoBaseRotations, OrthoViews, OrthoViewToNumber } from "viewer/constants";
+import { OrthoViews, OrthoViewToNumber } from "viewer/constants";
 import { getClosestHoveredBoundingBox } from "viewer/controller/combinations/bounding_box_handlers";
 import getSceneController from "viewer/controller/scene_controller_provider";
 import { getEnabledColorLayers } from "viewer/model/accessors/dataset_accessor";
@@ -48,6 +48,7 @@ import { showContextMenuAction } from "viewer/model/actions/ui_actions";
 import Dimensions from "viewer/model/dimensions";
 import { getBaseVoxelFactorsInUnit } from "viewer/model/scaleinfo";
 import type { Edge, Node, Tree } from "viewer/model/types/tree_types";
+import { OrthoBaseRotations } from "viewer/ortho_base_rotations";
 import { api } from "viewer/singletons";
 import Store from "viewer/store";
 import type FlightModeView from "viewer/view/arbitrary_view";
