+++ frontend/javascripts/viewer/model/sagas/skeletontracing_saga.ts
@@ -26,12 +26,7 @@ import {
   throttle,
 } from "typed-redux-saga";
 import { AnnotationLayerEnum, type ServerSkeletonTracing } from "types/api_types";
-import {
-  NumberToOrthoView,
-  OrthoBaseRotations,
-  TreeTypeEnum,
-  type Vector3,
-} from "viewer/constants";
+import { NumberToOrthoView, TreeTypeEnum, type Vector3 } from "viewer/constants";
 import { getSegmentIdForPositionAsync } from "viewer/controller/combinations/volume_handlers";
 import {
   isConcurrentCollaborationMode,
@@ -92,6 +87,7 @@ import {
   updateTreeGroupsExpandedState,
   updateTreeVisibility,
 } from "viewer/model/sagas/volume/update_actions";
+import { OrthoBaseRotations } from "viewer/ortho_base_rotations";
 import { api, Model } from "viewer/singletons";
 import type { SkeletonTracing, WebknossosState } from "viewer/store";
 import Store from "viewer/store";
